requirements.txt
annotated-types==0.7.0
anthropic==0.35.0
anyio==4.6.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
filelock==3.16.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
# huggingface-hub==0.25.1  # removed – unused heavy dependency
idna==3.10
Jinja2==3.1.4
# jiter==0.5.0  # removed – unused
# joblib==1.4.2  # removed – unused heavy dependency
MarkupSafe==2.1.5
# mpmath==1.3.0  # removed – unused heavy dependency
# networkx==3.3  # removed – unused
nltk==3.9.1
numpy==2.1.2
openai==1.51.0
packaging==24.1
# pandas==2.2.3  # removed – unused heavy dependency
# pillow==10.4.0  # removed – unused heavy dependency
# proto-plus==1.24.0  # removed – unused heavy dependency
# protobuf==5.28.2  # removed – unused heavy dependency
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.2.1
pydantic_core==2.23.4
pyparsing==3.1.4
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qasync==0.27.1
regex==2024.9.11
requests==2.32.3
# rsa==4.9  # removed – unused heavy dependency
# safetensors==0.4.5  # removed – unused heavy dependency
# scikit-learn==1.5.2  # removed – unused heavy dependency
# scipy==1.14.1  # removed – unused heavy dependency
# sentence-transformers==3.1.1  # removed – unused heavy dependency
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
# sympy==1.13.3  # removed – unused heavy dependency
# threadpoolctl==3.5.0  # removed – unused heavy dependency
tiktoken==0.8.0
# tokenizers==0.20.0  # removed – unused heavy dependency
# torch==2.4.1  # removed – unused heavy dependency
# tqdm==4.66.5  # removed – unused heavy dependency
# transformers==4.45.1  # removed – unused heavy dependency
typing_extensions==4.12.2
# tzdata==2024.2  # removed – unused heavy dependency
uritemplate==4.1.1
urllib3==2.2.3

# Development dependencies (optional)
# mypy==1.11.2
# types-requests
# types-python-dateutil
